rename shortener_core.py to shortener_utils.py
