+++ url_shortener/tests/test_shortener_storage.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 
 from ..models import ShortenedUrl
 from ..shortener import shortener_storage
-from ..shortener.shortener_core import convert_url_handle_to_number
+from ..shortener.shortener_utils import convert_url_handle_to_number
 
 
 class TestShortenerStorage(TestCase):
